Add result class for stations, add methods for querying data from a stations result
