+++ wetterdienst/core/scalar/values.py
@@ -2,131 +2,22 @@
 # Copyright (c) 2018-2021, earthobservations developers.
 # Distributed under the MIT License. See LICENSE for more info.
 from abc import abstractmethod
-from datetime import datetime
 from enum import Enum
-from logging import getLogger
-from typing import Dict, Generator, List, Optional, Tuple, Union
+from typing import Tuple, List, Union, Generator, Dict
+from dataclasses import dataclass
 
-import dateutil.parser
-import numpy as np
 import pandas as pd
-import pytz
 from pytz import timezone
 from tqdm import tqdm
 
-from wetterdienst.core.core import Core
-from wetterdienst.dwd.metadata.column_names import DWDMetaColumns
-from wetterdienst.exceptions import NoParametersFound, StartDateEndDateError
+from wetterdienst import Period
+from wetterdienst.core.scalar.core import ScalarCore
+from wetterdienst.core.scalar.result import StationsResult, ValuesResult
 from wetterdienst.metadata.columns import Columns
-from wetterdienst.metadata.period import Period, PeriodType
-from wetterdienst.metadata.resolution import Frequency, Resolution, ResolutionType
-from wetterdienst.metadata.result import Result
+from wetterdienst.metadata.period import PeriodType
+from wetterdienst.metadata.resolution import ResolutionType
 from wetterdienst.metadata.timezone import Timezone
 from wetterdienst.util.enumeration import parse_enumeration_from_template
-from wetterdienst.util.geo import Coordinates, derive_nearest_neighbours
-
-log = getLogger(__name__)
-
-
-EARTH_RADIUS_KM = 6371
-
-# TODO: move more attributes to __init__
-
-
-class ScalarCore(Core):
-    """Core for time series related classes """
-
-    @property
-    def resolution(self) -> Optional[Resolution]:
-        """ Resolution accessor"""
-        return self._resolution
-
-    @resolution.setter
-    def resolution(self, res) -> None:
-        # TODO: add functionality to parse arbitrary resolutions for cases where
-        #  resolution has to be determined based on returned data
-        if self._resolution_type in (ResolutionType.FIXED, ResolutionType.UNDEFINED):
-            self._resolution = res
-        else:
-            self._resolution = parse_enumeration_from_template(
-                res, self._resolution_base, Resolution
-            )
-
-    @property
-    @abstractmethod
-    def _resolution_base(self) -> Optional[Resolution]:
-        """ Optional enumeration for multiple resolutions """
-        pass
-
-    @property
-    @abstractmethod
-    def _resolution_type(self) -> ResolutionType:
-        """ Resolution type, multi, fixed, ..."""
-        pass
-
-    # TODO: implement for source with dynamic resolution
-    @staticmethod
-    def _determine_resolution(dates: pd.Series) -> Resolution:
-        """ Function to determine resolution from a pandas Series of dates """
-        pass
-
-    @property
-    def frequency(self) -> Frequency:
-        """Frequency for the given resolution, used to create a full date range for
-        mering"""
-        return Frequency[self.resolution.name]
-
-    @property
-    @abstractmethod
-    def _period_type(self) -> PeriodType:
-        """ Period type, fixed, multi, ..."""
-        pass
-
-    @property
-    @abstractmethod
-    def _period_base(self) -> Optional[Period]:
-        """ Period base enumeration from which a period string can be parsed """
-        pass
-
-    @abstractmethod
-    def _parse_period(self, period: List[Period]):
-        """ Method for parsing period depending on if multiple are given """
-        pass
-
-    def __init__(
-        self,
-        resolution: Resolution,
-        period: Union[Period, List[Period]],
-        start_date: Optional[Union[str, datetime]],
-        end_date: Optional[Union[str, datetime]],
-    ) -> None:
-        self.resolution = resolution
-        self.period = self._parse_period(period)
-
-        if start_date or end_date:
-            # If only one date given, set the other one to equal
-            if not start_date:
-                start_date = end_date
-
-            if not end_date:
-                end_date = start_date
-
-            start_date = dateutil.parser.isoparse(str(start_date))
-            if not start_date.tzinfo:
-                start_date = start_date.replace(tzinfo=pytz.UTC)
-
-            end_date = dateutil.parser.isoparse(str(end_date))
-            if not end_date.tzinfo:
-                end_date = end_date.replace(tzinfo=pytz.UTC)
-
-            # TODO: replace this with a response + logging
-            if not start_date <= end_date:
-                raise StartDateEndDateError(
-                    "Error: 'start_date' must be smaller or equal to 'end_date'."
-                )
-
-        self.start_date = start_date
-        self.end_date = end_date
 
 
 class ScalarValuesCore(ScalarCore):
@@ -222,48 +113,62 @@ class ScalarValuesCore(ScalarCore):
                 .tolist()
             )
 
+    # def __init__(
+    #     self,
+    #     station_id: Tuple[str],
+    #     parameter: Tuple[Union[str, Enum]],
+    #     resolution: Resolution,
+    #     period: Period,
+    #     start_date: Optional[Union[str, datetime]],
+    #     end_date: Optional[Union[str, datetime]],
+    #     humanize_parameters: bool,
+    #     tidy_data: bool,
+    # ) -> None:
+    #     """
+    #
+    #     :param station_id: station ids for which data is requested
+    #     :param parameter: parameters either as strings or enumerations for which data
+    #         is requested
+    #     :param start_date: start date of the resulting data,
+    #         if not start_date: start_date = end_date
+    #     :param end_date: end date of the resulting data
+    #         if not end_date: end_date = start_date
+    #     :param humanize_parameters: bool if parameters should be renamed to meaningful
+    #         names
+    #
+    #     """
+    #     super(ScalarValuesCore, self).__init__(
+    #         resolution=resolution,
+    #         period=period,
+    #         start_date=start_date,
+    #         end_date=end_date,
+    #     )
+    #
+    #     # Make sure we receive a list of ids
+    #     self.station_ids = pd.Series(station_id).astype(str).tolist()
+    #     self.parameters = self._parse_parameters(parameter)
+    #
+    #     # TODO: replace this with a response + logging
+    #     # TODO: move this to self.collect_data
+    #     if not self.parameters:
+    #         raise NoParametersFound(f"No parameters could be parsed from {parameter}")
+    #
+    #     self.humanize_parameters = humanize_parameters
+    #     self.tidy_data = tidy_data
+
     def __init__(
-        self,
-        station_id: Tuple[str],
-        parameter: Tuple[Union[str, Enum]],
-        resolution: Resolution,
-        period: Period,
-        start_date: Optional[Union[str, datetime]],
-        end_date: Optional[Union[str, datetime]],
-        humanize_parameters: bool,
-        tidy_data: bool,
+            self,
+            stations: StationsResult
     ) -> None:
-        """
+        self.stations = stations
 
-        :param station_id: station ids for which data is requested
-        :param parameter: parameters either as strings or enumerations for which data
-            is requested
-        :param start_date: start date of the resulting data,
-            if not start_date: start_date = end_date
-        :param end_date: end date of the resulting data
-            if not end_date: end_date = start_date
-        :param humanize_parameters: bool if parameters should be renamed to meaningful
-            names
+    @classmethod
+    def from_stations(cls, stations: StationsResult):
+        if not cls._source == stations._source:
+            # TODO: change to SourceError
+            raise Exception("sources don't match")
 
-        """
-        super(ScalarValuesCore, self).__init__(
-            resolution=resolution,
-            period=period,
-            start_date=start_date,
-            end_date=end_date,
-        )
-
-        # Make sure we receive a list of ids
-        self.station_ids = pd.Series(station_id).astype(str).tolist()
-        self.parameters = self._parse_parameters(parameter)
-
-        # TODO: replace this with a response + logging
-        # TODO: move this to self.collect_data
-        if not self.parameters:
-            raise NoParametersFound(f"No parameters could be parsed from {parameter}")
-
-        self.humanize_parameters = humanize_parameters
-        self.tidy_data = tidy_data
+        return cls(stations)
 
     def __eq__(self, other):
         """ Equal method of request object """
@@ -423,7 +328,7 @@ class ScalarValuesCore(ScalarCore):
                 continue
 
             # TODO: add meaningful metadata here
-            yield Result(pd.DataFrame(), station_df)
+            yield ValuesResult(pd.DataFrame(), station_df)
 
     @abstractmethod
     def _collect_station_parameter(self, station_id: str, parameter) -> pd.DataFrame:
@@ -614,191 +519,3 @@ class ScalarValuesCore(ScalarCore):
         hcnm = {parameter.value: parameter.name for parameter in self._parameter_base}
 
         return hcnm
-
-
-class ScalarStationsCore(ScalarCore):
-    """ Core for stations information of a source """
-
-    # Columns that should be contained within any stations information
-    _base_columns = (
-        Columns.STATION_ID.value,
-        Columns.FROM_DATE.value,
-        Columns.TO_DATE.value,
-        Columns.HEIGHT.value,
-        Columns.LATITUDE.value,
-        Columns.LONGITUDE.value,
-        Columns.STATION_NAME.value,
-        Columns.STATE.value,
-    )
-    # TODO: eventually this can be matched with the type coercion of station data to get
-    #  similar types of floats and strings
-    # Dtype mapping for stations
-    _dtype_mapping = {
-        Columns.STATION_ID.value: str,
-        Columns.HEIGHT.value: float,
-        Columns.LATITUDE.value: float,
-        Columns.LONGITUDE.value: float,
-        Columns.STATION_NAME.value: str,
-        Columns.STATE.value: str,
-    }
-
-    def _parse_period(self, period: Period):
-        if not period:
-            return None
-        elif self._period_type == PeriodType.FIXED:
-            return period
-        else:
-            return parse_enumeration_from_template(period, self._period_base, Period)
-
-    def __init__(
-        self,
-        resolution: Resolution,
-        period: Period,
-        start_date: Union[None, str, datetime] = None,
-        end_date: Union[None, str, datetime] = None,
-    ) -> None:
-        """
-
-        :param start_date: start date for filtering stations for their available data
-        :param end_date: end date for filtering stations for their available data
-        """
-        super(ScalarStationsCore, self).__init__(
-            resolution=resolution,
-            period=period,
-            start_date=start_date,
-            end_date=end_date,
-        )
-
-    def all(self) -> pd.DataFrame:
-        """
-        Wraps the _all method and applies date filters.
-
-        :return: pandas.DataFrame with the information of different available stations
-        """
-        metadata_df = self._all().copy()
-
-        metadata_df = metadata_df.reindex(columns=self._base_columns)
-
-        metadata_df = self._coerce_meta_fields(metadata_df)
-
-        if self.start_date:
-            metadata_df = metadata_df[
-                metadata_df[DWDMetaColumns.FROM_DATE.value] <= self.start_date
-            ]
-
-        if self.end_date:
-            metadata_df = metadata_df[
-                metadata_df[DWDMetaColumns.TO_DATE.value] >= self.end_date
-            ]
-
-        return metadata_df
-
-    def _coerce_meta_fields(self, df) -> pd.DataFrame:
-        """ Method for filed coercion. """
-        df = df.astype(self._dtype_mapping)
-
-        df[Columns.FROM_DATE.value] = pd.to_datetime(
-            df[Columns.FROM_DATE.value], infer_datetime_format=True
-        ).dt.tz_localize(pytz.UTC)
-        df[Columns.TO_DATE.value] = pd.to_datetime(
-            df[Columns.TO_DATE.value], infer_datetime_format=True
-        ).dt.tz_localize(pytz.UTC)
-
-        return df
-
-    @abstractmethod
-    def _all(self) -> pd.DataFrame:
-        """
-        Abstract method for gathering of sites information for a given implementation.
-        Information consist of a DataFrame with station ids, location, name, etc
-
-        :return: pandas.DataFrame with the information of different available sites
-        """
-        pass
-
-    def nearby_number(
-        self,
-        latitude: float,
-        longitude: float,
-        number: int,
-    ) -> pd.DataFrame:
-        """
-        Wrapper for get_nearby_stations_by_number using the given parameter set. Returns
-        nearest stations defined by number.
-
-        :param latitude: latitude in degrees
-        :param longitude: longitude in degrees
-        :param number: number of stations to be returned, greater 0
-        :return: pandas.DataFrame with station information for the selected stations
-        """
-        if number <= 0:
-            raise ValueError("'num_stations_nearby' has to be at least 1.")
-
-        coords = Coordinates(np.array(latitude), np.array(longitude))
-
-        metadata = self.all()
-
-        metadata = metadata.reset_index(drop=True)
-
-        distances, indices_nearest_neighbours = derive_nearest_neighbours(
-            metadata[Columns.LATITUDE.value].values,
-            metadata[Columns.LONGITUDE.value].values,
-            coords,
-            number,
-        )
-
-        distances = pd.Series(distances)
-        indices_nearest_neighbours = pd.Series(indices_nearest_neighbours)
-
-        # If num_stations_nearby is higher then the actual amount of stations
-        # further indices and distances are added which have to be filtered out
-        distances = distances[: min(metadata.shape[0], number)]
-        indices_nearest_neighbours = indices_nearest_neighbours[
-            : min(metadata.shape[0], number)
-        ]
-
-        distances_km = np.array(distances * EARTH_RADIUS_KM)
-
-        metadata_location = metadata.iloc[indices_nearest_neighbours, :].reset_index(
-            drop=True
-        )
-
-        metadata_location[DWDMetaColumns.DISTANCE_TO_LOCATION.value] = distances_km
-
-        if metadata_location.empty:
-            log.warning(
-                f"No weather stations were found for coordinate "
-                f"{latitude}°N and {longitude}°E "
-            )
-
-        return metadata_location
-
-    def nearby_radius(
-        self,
-        latitude: float,
-        longitude: float,
-        max_distance_in_km: int,
-    ) -> pd.DataFrame:
-        """
-        Wrapper for get_nearby_stations_by_distance using the given parameter set.
-        Returns nearest stations defined by distance (km).
-
-        :param latitude: latitude in degrees
-        :param longitude: longitude in degrees
-        :param max_distance_in_km: distance (km) for which stations will be selected
-        :return: pandas.DataFrame with station information for the selected stations
-        """
-        # Theoretically a distance of 0 km is possible
-        if max_distance_in_km < 0:
-            raise ValueError("'max_distance_in_km' has to be at least 0.0.")
-
-        metadata = self.all()
-
-        all_nearby_stations = self.nearby_number(latitude, longitude, metadata.shape[0])
-
-        nearby_stations_in_distance = all_nearby_stations[
-            all_nearby_stations[DWDMetaColumns.DISTANCE_TO_LOCATION.value]
-            <= max_distance_in_km
-        ]
-
-        return nearby_stations_in_distance.reset_index(drop=True)
